Implement blocks for two columns layout
