+++ store/interfaces.json
@@ -1,22 +1,43 @@
 {
   "checkout-cart": {
-    "component": "Cart",
+    "component": "CartWrapper",
     "allowed": [
-      "checkout-summary",
-      "empty-state",
-      "product-list",
-      "shipping-calculator"
+      "checkout-cart-two-cols",
+      "empty-state"
     ]
   },
   "empty-state": {
     "component": "vtex.render-runtime/LayoutContainer",
     "composition": "children",
     "allowed": "*"
-  },
+  },  
   "empty-state-icon": {
     "component": "EmptyStateIcon"
   },
   "empty-state-message": {
     "component": "EmptyStateMessage"
+  },
+  "checkout-cart-two-cols": {
+    "component": "vtex.render-runtime/LayoutContainer",
+    "composition": "children",
+    "allowed": "*"
+  },
+  "title": {
+    "component": "Title"
+  },
+  "product-list-wrapper": {
+    "component": "ProductList",
+    "allowed": ["product-list"]
+  },
+  "continue-shopping": {
+    "component": "ContinueShopping"
+  },
+  "shipping-wrapper": {
+    "component": "Shipping",
+    "allowed": ["shipping-calculator"]
+  },
+  "summary-wrapper": {
+    "component": "Summary",
+    "allowed": ["checkout-summary"]
   }
 }
